requirements-dev.txt
paddlepaddle>=2.4.1
pre-commit
pytest
parameterized
pytest-cov
regex
pytest-xdist
